COM-1960 - create 123paie folder
